move original search fragment to dialog search, create new search fragment
