+++ app/src/main/java/com/example/spaceshare/ui/view/SearchFragment.kt
@@ -1,30 +1,22 @@
 package com.example.spaceshare.ui.view
 
-import MapDialogFragment
 import android.location.Geocoder
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.view.isGone
 import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
 import androidx.navigation.NavController
-import androidx.navigation.findNavController
 import com.example.spaceshare.R
+import com.example.spaceshare.databinding.DialogSearchBinding
 import com.example.spaceshare.databinding.FragmentSearchBinding
 import com.example.spaceshare.ui.viewmodel.SearchViewModel
-import com.google.android.material.datepicker.CalendarConstraints
-import com.google.android.material.datepicker.DateValidatorPointForward
-import com.google.android.material.datepicker.MaterialDatePicker
 import dagger.hilt.android.AndroidEntryPoint
-import java.util.Objects
 import javax.inject.Inject
 
 @AndroidEntryPoint
 class SearchFragment : Fragment() {
-
     companion object {
         private val TAG = this::class.simpleName
     }
@@ -44,89 +36,23 @@ class SearchFragment : Fragment() {
         return binding.root
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        navController = requireActivity().findNavController(R.id.main_nav_host_fragment)
-        geocoder = Geocoder(requireContext())
-
-        configureCards()
-        configureButtons()
-    }
-
-    private fun configureCards() {
-        // Where
-        binding.whereCard.setOnClickListener {
-            hideWhatSelectorCard()
-            val mapDialogFragment = MapDialogFragment(searchViewModel)
-            mapDialogFragment.show(Objects.requireNonNull(childFragmentManager), "mapDialog")
-        }
-        searchViewModel.location?.observe(viewLifecycleOwner) { location ->
-            if (location != null) {
-                val addresses = geocoder.getFromLocation(location.latitude, location.longitude, 1)
-                if (!addresses.isNullOrEmpty()) {
-                    val address = addresses[0]
-                    binding.searchLocation.text = address.getAddressLine(0)
+    companion object {
+        /**
+         * Use this factory method to create a new instance of
+         * this fragment using the provided parameters.
+         *
+         * @param param1 Parameter 1.
+         * @param param2 Parameter 2.
+         * @return A new instance of fragment SearchFragment.
+         */
+        // TODO: Rename and change types and number of parameters
+        @JvmStatic
+        fun newInstance(param1: String, param2: String) =
+            SearchFragment().apply {
+                arguments = Bundle().apply {
+                    putString(ARG_PARAM1, param1)
+                    putString(ARG_PARAM2, param2)
                 }
-            } else {
-                binding.searchLocation.text = "Anywhere"
             }
-        }
-
-        // When
-        val constraintsBuilder = CalendarConstraints.Builder()
-            .setValidator(DateValidatorPointForward.now())
-        val dateRangePicker =
-            MaterialDatePicker.Builder.dateRangePicker()
-                .setTitleText("Select Dates")
-                .setCalendarConstraints(constraintsBuilder.build())
-                .build()
-        dateRangePicker.addOnPositiveButtonClickListener {
-            binding.searchTime.text = dateRangePicker.headerText
-            searchViewModel.startTime.value = dateRangePicker.selection?.first
-            searchViewModel.endTime.value = dateRangePicker.selection?.second
-        }
-        binding.whenCard.setOnClickListener {
-            dateRangePicker.show(parentFragmentManager, TAG)
-            hideWhatSelectorCard()
-        }
-        searchViewModel.endTime.observe(viewLifecycleOwner) {
-            if (it == System.currentTimeMillis())
-                binding.searchTime.text = "Anytime"
-        }
-
-        // What
-        binding.whatCard.setOnClickListener {
-            it.isGone = true
-            binding.whatSelectorCard.isGone = false
-        }
-        binding.whatSelectorAddSize.setOnClickListener {
-            searchViewModel.incrementSpaceRequired()
-        }
-        binding.whatSelectorMinusSize.setOnClickListener {
-            searchViewModel.decrementSpaceRequired()
-        }
-        searchViewModel.spaceRequired.observe(viewLifecycleOwner) { spaceRequired ->
-            binding.whatSelectorSizeText.text = spaceRequired.toString()
-            if (spaceRequired == 0.0)
-                binding.searchSize.text = "Anysize"
-            else
-                binding.searchSize.text = spaceRequired.toString() + " cubic metres"
-        }
     }
-
-    private fun configureButtons() {
-        binding.btnClear.setOnClickListener {
-            searchViewModel.clearAllData()
-        }
-
-        binding.btnSearch.setOnClickListener {
-            searchViewModel.submitSearch()
-        }
-    }
-
-    private fun hideWhatSelectorCard() {
-        binding.whatCard.isGone = false
-        binding.whatSelectorCard.isGone = true
-    }
-
 }
